+++ domain/src/main/java/com/papmobdev/domain/diagnostic/usecasesensorevents/SendEventModel.kt
@@ -3,6 +3,6 @@ package com.papmobdev.domain.diagnostic.usecasesensorevents
 import com.papmobdev.domain.sensor.models.EventModel
 
 data class SendEventModel(
-    val idDiagnostic: Int,
+    val idDiagnostic: Long,
     val listEvents: List<EventModel>
 )
